requirements.txt
openai>=1.12.0
requests>=2.31.0
moviepy==1.0.3
Pillow>=10.0.0
numpy>=1.24.0
srt>=3.5.3
imageio-ffmpeg>=0.4.9
